// src/api/generated/uberiket/models/inkluder-iveglenkesekvenser.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NVDB API Les Uberiket
 * Dette API-et brukes til å lese uberikede hendelser, vegobjekter og vegnett.<br><br>[Dokumentasjon for Uberiket API](https://nvdb-docs.atlas.vegvesen.no/nvdbapil/v4/Uberiket)
 *
 * The version of the OpenAPI document: v1
 * Contact: nvdb@vegvesen.no
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export const InkluderIVeglenkesekvenser = {
    Minimum: 'minimum',
    Alle: 'alle'
} as const;

export type InkluderIVeglenkesekvenser = typeof InkluderIVeglenkesekvenser[keyof typeof InkluderIVeglenkesekvenser];