// src/api/generated/uberiket/models/inkluder-ivegobjekt.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NVDB API Les Uberiket
 * Dette API-et brukes til å lese uberikede hendelser, vegobjekter og vegnett.<br><br>[Dokumentasjon for Uberiket API](https://nvdb-docs.atlas.vegvesen.no/nvdbapil/v4/Uberiket)
 *
 * The version of the OpenAPI document: v1
 * Contact: nvdb@vegvesen.no
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export const InkluderIVegobjekt = {
    Gyldighetsperiode: 'gyldighetsperiode',
    Egenskaper: 'egenskaper',
    Stedfesting: 'stedfesting',
    Barn: 'barn',
    Alle: 'alle',
    Minimum: 'minimum'
} as const;

export type InkluderIVegobjekt = typeof InkluderIVegobjekt[keyof typeof InkluderIVegobjekt];