// src/api/generated/uberiket/models/vegobjekt.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NVDB API Les Uberiket
 * Dette API-et brukes til å lese uberikede hendelser, vegobjekter og vegnett.<br><br>[Dokumentasjon for Uberiket API](https://nvdb-docs.atlas.vegvesen.no/nvdbapil/v4/Uberiket)
 *
 * The version of the OpenAPI document: v1
 * Contact: nvdb@vegvesen.no
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EgenskapVerdi } from './egenskap-verdi';
// May contain unused imports in some cases
// @ts-ignore
import type { Gyldighetsperiode } from './gyldighetsperiode';
// May contain unused imports in some cases
// @ts-ignore
import type { Stedfesting } from './stedfesting';

/**
 * 
 * @export
 * @interface Vegobjekt
 */
export interface Vegobjekt {
    /**
     * 
     * @type {number}
     * @memberof Vegobjekt
     */
    'id': number;
    /**
     * 
     * @type {number}
     * @memberof Vegobjekt
     */
    'versjon': number;
    /**
     * 
     * @type {number}
     * @memberof Vegobjekt
     */
    'typeId': number;
    /**
     * Tilstede dersom inkluder=alle|gyldighetsperiode
     * @type {Gyldighetsperiode}
     * @memberof Vegobjekt
     */
    'gyldighetsperiode'?: Gyldighetsperiode;
    /**
     * Tilstede dersom inkluder=alle|egenskaper
     * @type {{ [key: string]: EgenskapVerdi; }}
     * @memberof Vegobjekt
     */
    'egenskaper'?: { [key: string]: EgenskapVerdi; };
    /**
     * Tilstede dersom inkluder=alle|barn
     * @type {{ [key: string]: Set<number>; }}
     * @memberof Vegobjekt
     */
    'barn'?: { [key: string]: Set<number>; };
    /**
     * Tilstede dersom inkluder=alle|stedfesting og objektet er stedfestet på vegnettet
     * @type {Stedfesting}
     * @memberof Vegobjekt
     */
    'stedfesting'?: Stedfesting;
    /**
     * 
     * @type {string}
     * @memberof Vegobjekt
     */
    'sistEndret': string;
}