// src/api/generated/uberiket/models/stedfesting-linje.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NVDB API Les Uberiket
 * Dette API-et brukes til å lese uberikede hendelser, vegobjekter og vegnett.<br><br>[Dokumentasjon for Uberiket API](https://nvdb-docs.atlas.vegvesen.no/nvdbapil/v4/Uberiket)
 *
 * The version of the OpenAPI document: v1
 * Contact: nvdb@vegvesen.no
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Retning } from './retning';
// May contain unused imports in some cases
// @ts-ignore
import type { Sideposisjon } from './sideposisjon';

/**
 * 
 * @export
 * @interface StedfestingLinje
 */
export interface StedfestingLinje {
    /**
     * 
     * @type {number}
     * @memberof StedfestingLinje
     */
    'id': number;
    /**
     * 
     * @type {number}
     * @memberof StedfestingLinje
     */
    'startposisjon': number;
    /**
     * 
     * @type {number}
     * @memberof StedfestingLinje
     */
    'sluttposisjon': number;
    /**
     * 
     * @type {Retning}
     * @memberof StedfestingLinje
     */
    'retning'?: Retning;
    /**
     * 
     * @type {Sideposisjon}
     * @memberof StedfestingLinje
     */
    'sideposisjon'?: Sideposisjon;
    /**
     * 
     * @type {Array<string>}
     * @memberof StedfestingLinje
     */
    'kjorefelt': Array<string>;
}